requirements.txt
# /home/user/document_classifier/requirements.txt

Flask==2.2.2
gunicorn==20.1.0
pandas==1.5.2
scikit-learn==1.2.0
tika-python==2.6.0
langdetect==1.0.9